Serializers, models, and url confs
Serializers for all endpoints
Models for financial_activities
Refactored some misplaced models
Updated URL confs
