+++ usaspending_api/accounts/serializers.py
@@ -8,7 +8,7 @@ class TreasuryAppropriationAccountSerializer(serializers.ModelSerializer):
     class Meta:
 
         model = TreasuryAppropriationAccount
-        fields = ('treasury_account_identifier', 'allocation_transfer_agency_id')
+        fields = '__all__'
 
 
 class AppropriationAccountBalancesSerializer(serializers.ModelSerializer):
@@ -16,4 +16,4 @@ class AppropriationAccountBalancesSerializer(serializers.ModelSerializer):
     class Meta:
 
         model = AppropriationAccountBalances
-        fields = ('budget_authority_unobligat_fyb', 'budget_authority_appropria_cpe')
+        fields = '__all__'
