+++ usaspending_api/accounts/urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import url
+from usaspending_api.accounts import views
+
+urlpatterns = [
+    url(r'^', views.AppropriationAccountBalancesList.as_view()),
+]
