+++ usaspending_api/financial_activities/urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import url
+from usaspending_api.financial_activities import views
+
+urlpatterns = [
+    url(r'^', views.FinancialAccountsByProgramActivityObjectClassList.as_view()),
+]
