+++ usaspending_api/submissions/urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import url
+from usaspending_api.submissions import views
+
+urlpatterns = [
+    url(r'^', views.SubmissionAttributesList.as_view()),
+]
