+++ usaspending_api/urls.py
@@ -20,5 +20,8 @@ from django.contrib import admin
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url(r'^api/v1/awards/', include('usaspending_api.awards.urls')),
+    url(r'^api/v1/submissions/', include('usaspending_api.submissions.urls')),
+    url(r'^api/v1/accounts/', include('usaspending_api.accounts.urls')),
+    url(r'^api/v1/financial_activities/', include('usaspending_api.financial_activities.urls')),
     url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
 ]
